+++ interfaces/index.ts
@@ -9,3 +9,4 @@ export * from './bank-account.interface'
 export * from './e-wallet.interface'
 export * from './member.interface'
 export * from './seller.interface'
+export * from './vocher.interface'
