+++ modules/SellerMarketing/components/Voucher/components/AddVocher/AddVocher.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../public/sass/config/variables';
+@import '../../../../../../public/sass/config/variables';
 
 .textSecondary {
   color: $secondary-color !important;
