+++ modules/SellerMarketing/components/Voucher/components/VoucherList/VoucherList.module.scss
@@ -1,0 +1,28 @@
+@import '../../../../../../public/sass/config/variables';
+
+.textPrimary {
+  color: $primary-color;
+}
+
+.textSecondary {
+  color: $secondary-color !important;
+}
+
+.highlight {
+  padding: 1.5rem;
+  margin-bottom: 1.5rem;
+  background-color: $background-light;
+  border-radius: $border-radius;
+
+  @media (min-width: 768px) {
+    padding: 1.5rem 0;
+  }
+}
+
+.hrTitleCol {
+  border-bottom: 1px solid red;
+}
+
+.tr {
+  border-bottom: 1px solid #d9d9d9;
+}
