+++ pages/seller/settings/marketing/add-voucher.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react'
 import { NextPageContext } from 'next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
-import VoucherForm from '~/modules/SellerMarketing/components/VoucherForm'
+import AddVocher from '~/modules/SellerMarketing/components/Voucher/components/AddVocher'
 import { LocaleNamespaceConst } from '~/constants'
 
 export async function getServerSideProps(context: NextPageContext): Promise<any> {
@@ -14,6 +14,6 @@ export async function getServerSideProps(context: NextPageContext): Promise<any>
     }
   }
 }
-const VoucherAddPage: FC = () => <VoucherForm />
+const VoucherAddPage: FC = () => <AddVocher />
 
 export default VoucherAddPage
